src/graphql/traveldata/resolvers.ts
const AWS = require('aws-sdk');
const athena = new AWS.Athena();

const resolvers = {
    Query: {
        reservations: async (_, { limit = 10 }) => {
            const params = {
                QueryString: `SELECT * FROM your_table_name LIMIT ${limit}`,
                QueryExecutionContext: {
                    Database: 'athena-app-db'
                },
                ResultConfiguration: {
                    OutputLocation: 's3://your-output-bucket/path/'
                }
            };

            const results = await athena.startQueryExecution(params).promise();
            // Handle Athena results, transform them into the format expected by your GraphQL schema, and return.
        }
    }
};
